pyproject.toml
[project]
name = "mini-whisper-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-whisper>=20250625",
    "pyautogui>=0.9.54",
    "pynput>=1.8.1",
    "sounddevice>=0.5.2",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
]
